[Handler] Integrate the watcher's cache into request handler (#78)
### Description  
This PR integrates the cache into the handler in order to resolve the
request target based on the new cache.

### Changes Made  
- Add the `IngressHostCacheReader` (the cache read-only access) to the
handler
- Add the logic for the new resolving (i.e. - try to resolve based on
the cache, and if failed - resolve by the existing headers' logic)

### Testing  
- Added a flow test - for the entire request handling  
- Added unit tests for the resolving per case (i.e. - based on headers
and based on ingress)

### References  
- Jira ticket link: https://iguazio.atlassian.net/browse/NUC-523  
- External link: this PR is based on this branch -
https://github.com/v3io/scaler/pull/76

### Additional Notes  
- I left a **TODO** question inside the test file — it seems that when
parsing the `targetURL`, the handler appends the path twice. This fix is
relatively minor and has been tested manually, but since the current
code still works, it's worth deciding whether we want to address it now.